features/features_gfa_cts/requirements.txt
alabaster==0.7.12
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bcrypt==3.2.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
configobj==5.0.6
cryptography==3.4.7
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.3
dlhub-client==0.0.1
dlhub-sdk==0.9.4
docutils==0.16
entrypoints==0.3
fair-research-login==0.2.0
foundry-ml==0.0.7
funcx==0.0.5
future==0.18.2
globus-nexus-client==0.3.0
globus-sdk==2.0.1
h5py==3.2.1
idna==2.10
imagesize==1.2.0
ipykernel==5.5.3
ipyparallel==6.3.0
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
llvmlite==0.36.0
lockfile==0.12.2
MarkupSafe==1.1.1
matplotlib==3.4.1
mdf-connect-client==0.3.8
mdf-forge==0.7.6
mdf-toolbox==0.5.7
mistune==0.8.4
monty==2021.3.3
mpmath==1.2.1
nameparser==1.0.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nbsphinx==0.8.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.3.0
numba==0.53.1
numpy==1.20.2
packaging==20.9
palettable==3.3.0
pandas==1.2.4
pandocfilters==1.4.3
paramiko==2.7.2
parsl==0.9.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
plotly==4.14.3
prometheus-client==0.10.1
prompt-toolkit==3.0.18
psutil==5.8.0
ptyprocess==0.7.0
pycparser==2.20
pydantic==1.8.1
Pygments==2.8.1
PyJWT==1.7.1
pymatgen==2022.0.6
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-daemon==2.3.0
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.3
requests==2.25.1
retrying==1.3.3
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
scikit-learn==0.24.1
scipy==1.6.2
Send2Trash==1.5.0
six==1.15.0
snowballstemmer==2.1.0
spglib==1.16.1
Sphinx==3.5.4
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sympy==1.8
tabulate==0.8.9
tblib==1.7.0
terminado==0.9.4
testpath==0.4.4
texttable==1.6.3
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typeguard==2.12.0
typer==0.3.2
typing-extensions==3.7.4.3
uncertainties==3.1.5
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
